requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cmasher==1.6.3
colorama==0.4.6
colorspacious==1.1.2
contourpy==1.1.0
copier==8.1.0
cryptography==41.0.3
cycler==0.11.0
decorator==5.1.1
dunamai==1.18.0
e13tools==0.9.6
Envelopes==0.4
flet==0.8.1
flet-core==0.8.1
flet-runtime==0.8.1
fonttools==4.41.0
frozenlist==1.3.3
func-timeout==4.3.5
funcy==2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
kiwisolver==1.4.4
logging-utilities==4.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
multidict==6.0.4
mysql-connector-python==8.0.33
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.6
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
Pillow==10.0.0
plumbum==1.8.2
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pyAesCrypt==6.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
pyparsing==3.0.9
pypng==0.20220715.0
Pypubsub==4.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==1.10.0
repath==0.9.0
replicate==0.9.0
requests==2.30.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2